+++ admin/includes/view-users.php
@@ -33,7 +33,7 @@
                                        echo "<td>$user_email</td>";     
                                        echo "<td><img src='../img/$user_image' height ='45' width= '45'></td>";
                                                 
-                                       echo "<td><a class='btn btn-danger' href='users.php?delete=$user_id'>Delete</a></td>"; 
+                                       echo "<td><a class='btn btn-danger' onClick=\"return confirm('Da li ste sigurni?')\" href='users.php?delete=$user_id'>Delete</a></td>"; 
                                        echo "<td><a class='btn btn-primary' href='users.php?source=edit_user&user_id=$user_id'>Edit</a></td>"; 
                                          
                                     echo "</tr>";
